% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thorn03.R
\name{thorn03}
\alias{thorn03}
\title{Makes a thorns image}
\usage{
thorn03(
  shade,
  file = NULL,
  cells = 20000,
  layers = 5,
  threshold = 0.02,
  trim = 1
)
}
\arguments{
\item{shade}{A string specifying a hex colour (e.g., "#ff45ab")}

\item{file}{Name of the file to be created (if NULL, default is created using shade value)}

\item{cells}{Number of points used in the voronoi tesselation (default = 20000)}

\item{layers}{Number of function variants used in the iterated function system (default = 5)}

\item{threshold}{Maximum size for brightly coloured tiles (default = .02)}

\item{trim}{Threshold for discarding outlier points (default = 1)}
}
\description{
Makes a thorns image
}
